♻️refactor: 클래스의 책임 분리
1. Interceptor, RedisPublisher
As-Was: RedisPublisher에서 서비스로직 (입장 / 퇴장 메시지 작성)과 메시지 전송 로직이 혼재
AS-Is: publish 하는 책임만을 남기고 나니 redisTemplate.convertAndSend() 메서드 한 줄만 남아 아에 해당 클래스를 제거했음.
IntegerCeptor는 일종의 컨트롤러임. 사용자의 요청을 받아서 적절한 서비스 로직에 위임하고 다시 전달하는 책임. 이에 따라 리팩터링 진행함.

2. AuthUtil, AuthService
As-Was 인증 작업을 AuthUtil에서 진행하고 있었음.
As-Is AuthService를 만들어 책임을 분리함.
