+++ OneCoin/Server/src/main/java/OneCoin/Server/config/WebSocketConfig.java
@@ -1,13 +1,17 @@
 package OneCoin.Server.config;
 
-import OneCoin.Server.chat.publisher.RedisPublisher;
+import OneCoin.Server.chat.controller.MessageInterceptor;
+import OneCoin.Server.chat.mapper.ChatMapper;
 import OneCoin.Server.chat.service.ChatRoomService;
-import OneCoin.Server.chat.intercepter.MessageInterceptor;
+import OneCoin.Server.chat.service.ChatService;
+import OneCoin.Server.config.webSocketAuth.WebSocketAuthService;
 import OneCoin.Server.config.auth.utils.UserUtilsForWebSocket;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
@@ -22,7 +26,11 @@ import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerCo
 public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
     private final UserUtilsForWebSocket userUtilsForWebSocket;
     private final ChatRoomService chatRoomService;
-    private final RedisPublisher redisPublisher;
+    private final WebSocketAuthService webSocketAuthService;
+    private final ChatService chatService;
+    private final ChatMapper chatMapper;
+    private final RedisTemplate<Object, Object> redisTemplate;
+    private final ChannelTopic channelTopic;
 
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
@@ -47,7 +55,8 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
     @Override
     public void configureClientInboundChannel(ChannelRegistration registration) {
         registration.interceptors(new MessageInterceptor(
-                userUtilsForWebSocket, chatRoomService, redisPublisher));
+                userUtilsForWebSocket, chatRoomService, webSocketAuthService,
+                chatService, chatMapper, redisTemplate, channelTopic));
     }
