+++ OneCoin/Server/src/main/java/OneCoin/Server/config/auth/utils/UserUtilsForWebSocket.java
@@ -18,8 +18,6 @@ import java.util.Map;
 @AllArgsConstructor
 public class UserUtilsForWebSocket {
     private UserService userService;
-    private final JwtTokenizer jwtTokenizer;
-    private final CustomAuthorityUtils customAuthorityUtils;
 
     public User extractUser(Principal principal) {
         try {
@@ -40,25 +38,6 @@ public class UserUtilsForWebSocket {
         } catch (Exception e) {
             return null;
         }
-    }
-
-    public Authentication authenticate(String accessToken) {
-        Map<String, Object> claims = verifyJws(accessToken);
-        return getAuthentication(claims);
-    }
-
-    private Map<String, Object> verifyJws(String accessToken) {
-        String jws = accessToken.replace("Bearer ", "");
-        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey()); // SecretKey 파싱
-        Map<String, Object> claims = jwtTokenizer.getClaims(jws, base64EncodedSecretKey).getBody();   // Claims 파싱
-        return claims;
-    }
-
-    private Authentication getAuthentication(Map<String, Object> claims) {
-        List<GrantedAuthority> authorities = customAuthorityUtils.createAuthorities(Role.valueOf((String) claims.get("roles")));
-        Authentication authentication = new UsernamePasswordAuthenticationToken(claims, null, authorities);
-//        SecurityContextHolder.getContext().setAuthentication(authentication); // SecurityContext 에 Authentication 저장
-        return authentication;
     }
 
 }
